Fix case where target process name is not found
In some cases, /proc/pid/comm is not available to be parsed. If this
happens, return (null) string to show that.
