+++ tests/libstress.c
@@ -1,0 +1,4 @@
+int value(void)
+{
+  return 1;
+}
