+++ tests/libstress_livepatch1.c
@@ -1,0 +1,4 @@
+int new_value(void)
+{
+return 2;
+}
