+++ tests/manyprocesses.py
@@ -18,39 +18,6 @@
 import testsuite
 import subprocess
 
-def childless_livepatch(wildcard, timeout=10, retries=1,
-            verbose=False, quiet=False, revert_lib=None):
-
-    # Build command-line from arguments
-    command = [testsuite.ulptool, "trigger"]
-    if revert_lib is not None:
-      command.append("--revert-all")
-      command.append(revert_lib)
-    if wildcard is not None:
-      command.append(wildcard)
-      print('')
-    if verbose:
-      command.append('-v')
-    if quiet:
-      command.append('-q')
-    if retries > 1:
-      command.append('-r')
-      command.append(str(retries))
-
-    # Apply the live patch and check for common errors
-    try:
-      print('Applying/reverting live patch.')
-      tool = subprocess.run(command, timeout=timeout)
-    except subprocess.TimeoutExpired:
-      print('Live patching timed out.')
-      raise
-
-    # The trigger tool returns 0 on success, so use check_returncode(),
-    # which asserts that, and raises CalledProcessError otherwise.
-    tool.check_returncode()
-
-    print('Live patch applied/reverted successfully.')
-
 
 childs = [testsuite.spawn('manyprocesses'),
           testsuite.spawn('manyprocesses'),
@@ -65,7 +32,7 @@ for child in childs:
     child.expect('1-2-3-4-5-6-7-8');
     child.expect('1.0-2.0-3.0-4.0-5.0-6.0-7.0-8.0-9.0-10.0');
 
-childless_livepatch(wildcard='.libs/libmanyprocesses_livepatch1.so', verbose=True, revert_lib='libmanyprocesses.so.0')
+testsuite.childless_livepatch(wildcard='.libs/libmanyprocesses_livepatch1.so', verbose=True, revert_lib='libmanyprocesses.so.0')
 
 for child in childs:
     child.sendline('')
@@ -73,7 +40,7 @@ for child in childs:
     child.expect('10.0-9.0-8.0-7.0-6.0-5.0-4.0-3.0-2.0-1.0',
                  reject='1.0-2.0-3.0-4.0-5.0-6.0-7.0-8.0-9.0-10.0');
 
-childless_livepatch(wildcard=None, verbose=True, revert_lib='libmanyprocesses.so.0')
+testsuite.childless_livepatch(wildcard=None, verbose=True, revert_lib='libmanyprocesses.so.0')
 
 for child in childs:
     child.sendline('')
